Move UIEventArgs types
