+++ src/Components/Web/src/BindAttributes.cs
@@ -7,10 +7,10 @@ namespace Microsoft.AspNetCore.Components
     /// Infrastructure for the discovery of <c>bind</c> attributes for markup elements.
     /// </summary>
     /// <remarks>
-    /// To extend the set of <c>bind</c> attributes, define a public class named 
+    /// To extend the set of <c>bind</c> attributes, define a public class named
     /// <c>BindAttributes</c> and annotate it with the appropriate attributes.
     /// </remarks>
-    
+
     // Handles cases like <input @bind="..." /> - this is a fallback and will be ignored
     // when a specific type attribute is applied.
     [BindInputElement(null, null, "value", "onchange", isInvariantCulture: false, format: null)]
