+++ src/Components/Web/src/UIFocusEventArgs.cs
@@ -1,0 +1,14 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Components
+{
+    /// <summary>
+    /// Supplies information about a focus event that is being raised.
+    /// </summary>
+    public class UIFocusEventArgs : UIEventArgs
+    {
+        // Not including support for 'relatedTarget' since we don't have a good way to represent it.
+        // see: https://developer.mozilla.org/en-US/docs/Web/API/FocusEvent
+    }
+}
